refactor: creating user entity
