+++ firebase/functions/src/auth/index.ts
@@ -1,2 +1,1 @@
 export { signup } from "./handlers/signup";
-export { getUser } from "./handlers/getUser";
