+++ firebase/functions/src/user/handlers/getUser.ts
@@ -1,14 +1,14 @@
-import { HttpsError, onCall } from "firebase-functions/https";
+import { onCall } from "firebase-functions/https";
 import { withCallHandler } from "../../shared/middlewares/errorHandler";
 import { UserProfile } from "../types";
 import { requireAuthenticatedUser } from "../../shared/auth";
-import { getUserById } from "../repositories/userRepository";
+import { UserService } from "../shared/userService";
+
+const userService = new UserService();
 
 export const getUser = onCall(
   withCallHandler<void, UserProfile>(async (request) => {
     const uid = requireAuthenticatedUser(request).uid;
-    const user = await getUserById(uid);
-    if (!user) throw new HttpsError("not-found", "Usuário não encontrado");
-    return user;
+    return await userService.get(uid);
   }),
 );
