+++ firebase/functions/src/user/index.ts
@@ -1,0 +1,2 @@
+
+export { getUser } from "./handlers/getUser";
