+++ firebase/functions/src/user/types/index.ts
@@ -1,10 +1,5 @@
-import { FieldValue, Timestamp } from "firebase-admin/firestore";
-
-//
-// =========================
-// PERSISTIDO NO FIRESTORE
-// =========================
-//
+import { Timestamp } from "firebase-admin/firestore";
+import { WalletDTO, WalletTokenPositionDTO } from "./dtos";
 
 export type WalletTokenPosition = {
   startupId: string;
@@ -29,21 +24,7 @@ export type Wallet = {
   updatedAt: Timestamp;
 };
 
-export type WalletTokenPositionDTO = WalletTokenPosition & {
-  currentTokenPriceCents: number;
 
-  currentValueCents: number;
-};
-
-export type WalletDTO = Omit<Wallet, "positions"> & {
-  positions: WalletTokenPositionDTO[];
-};
-
-//
-// =========================
-// USER
-// =========================
-//
 
 export type UserProfile = {
   name: string;
@@ -56,9 +37,7 @@ export type UserProfile = {
   createdAt: Timestamp;
 };
 
-export type UserProfileDTO = Omit<UserProfile, "wallet"> & {
-  wallet: WalletDTO;
-};
+
 
 export type UpdateWalletParams = {
   userId: string;
@@ -72,10 +51,7 @@ export type UpdateWalletParams = {
   currentTokenPriceCents: number;
 };
 
-export type UserCreateDTO = UserProfile & {
-  uid: string;
-  createdAt: FieldValue;
-};
+
 
 export type UserEntity = UserProfile & {
   createdAt: Timestamp;
